(WIP) Move simulation to separate module.
